app.py

+++ app.py
@@ -53,7 +53,7 @@ import time
 #embedding_model_name = "sentence-transformers/all-MiniLM-L6-v2"
 embedding_model_name ="BAAI/bge-base-en-v1.5"
 
-model_sentence = SentenceTransformer(embedding_model_name, device="cpu", torch_dtype=torch.float32 )
+model_sentence = SentenceTransformer(embedding_model_name, device="cpu" )
 
 #compression and reranker model of the embedding
 cross_encoder_model = HuggingFaceCrossEncoder(model_name="BAAI/bge-reranker-base",model_kwargs={"device": "cpu"})
@@ -384,6 +384,7 @@ st.subheader(f"Recent Cyber Attacks and Breaches 🛡️")
 for i, title in enumerate(titles[:30],1):
     st.write(f"🔴({i})--> {title}")
             
+
